코드 살짝 다듬고, 삭제 컨펌 메시지 추가

+++ frontend/src/components/Items.js
@@ -22,10 +22,12 @@ const useStyles = makeStyles(theme => ({
 }));
 
 function List(props) {
-    const classes = useStyles();
+	const classes = useStyles();
     
     const handleRemove = () => {
-        props.onRemove(props.todo.id);
+		if(window.confirm('삭제할까요?')){
+            props.onRemove(props.todo.id);
+        }
     };
 
 	return (
